READ_ATTRIBUTE: Fix compiler warning on type comparison
In file included from vtltape.c:76:
vtltape.c: In function ‘resp_read_attribute’:
vtllib.h:50:21: warning: comparison of distinct pointer types lacks a cast
   50 |         (void) (&_x == &_y);    \
      |                     ^~
vtltape.c:705:16: note: in expansion of macro ‘min’
  705 |         return min(ret_val + 4, alloc_len);
      |                ^~~

Make variable types the same rather than explicit casting.

+++ usr/vtltape.c
@@ -643,7 +643,7 @@ int resp_read_attribute(struct scsi_cmd *cmd)
 {
 	uint16_t attrib;
 	uint32_t alloc_len;
-	int ret_val = 0;
+	uint32_t ret_val = 0;
 	int byte_index = 4;
 	int indx, found_attribute;
 	uint8_t *cdb = cmd->scb;
@@ -667,7 +667,7 @@ int resp_read_attribute(struct scsi_cmd *cmd)
 			if (found_attribute) {
 				/* calculate available data length */
 				ret_val += MAM_Attributes[indx].length + 5;
-				if ((uint32_t)ret_val < alloc_len) {
+				if (ret_val < alloc_len) {
 					/* add it to output */
 					buf[byte_index++] = MAM_Attributes[indx].attribute >> 8;
 					buf[byte_index++] = MAM_Attributes[indx].attribute;
@@ -691,7 +691,7 @@ int resp_read_attribute(struct scsi_cmd *cmd)
 		for (indx = found_attribute = 0; MAM_Attributes[indx].length; indx++) {
 			/* calculate available data length */
 			ret_val += 2;
-			if ((uint32_t)ret_val <= alloc_len) {
+			if (ret_val <= alloc_len) {
 				/* add it to output */
 				buf[byte_index++] = MAM_Attributes[indx].attribute >> 8;
 				buf[byte_index++] = MAM_Attributes[indx].attribute;
